python-agent/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
browser-use
langchain-openai
playwright
